switched the vps host workerman script over to utilizing a globalData settings variable to handle the various settings and switched a bunch of the code to implementing it

+++ workerman/start.php
@@ -7,10 +7,29 @@ use \Workerman\Lib\Timer;
 
 require_once __DIR__.'/vendor/autoload.php';
 
-define('CMD', 'htop'); // Command. For example 'tail -f /var/log/nginx/access.log'.
-define('HEARTBEAT_TIME', 600);
-define('ALLOW_CLIENT_INPUT', true); // Whether to allow client input.
-
+$settings = [
+	'auth' => [
+		'enable' => FALSE,
+		'timeout' => 30,
+	],
+	'vmstat' => [
+		'enable' => FALSE,
+	],
+	'phptty' => [
+		'enable' => FALSE,
+		'cmd' => 'htop', // Command. For example 'tail -f /var/log/nginx/access.log'.
+		'client_input' => TRUE, // Whether to allow client input.
+	],
+	'heartbeat' => [
+		'enable' => FALSE,
+		'check_interval' => 60,
+		'timeout' => 600,
+	],
+	'timers' => [
+		'update_vps_list' => 600,
+		'vps_queue' => 60,
+	],
+];
 
 function update_vps_list_timer() {
 	$task_connection = new AsyncTcpConnection('Text://127.0.0.1:55552');								// Asynchronous link with the remote task service
@@ -56,27 +75,43 @@ $worker = new Worker('Websocket://0.0.0.0:55554');
 $worker->name = 'WebsocketServer';
 
 // start the process, open a vmstat process, and broadcast vmstat process output to all browser clients
-$worker->onWorkerStart = function($worker) {
+$worker->onWorkerStart = function($worker) use ($settings) {
 	global $global;
 	$global = new \GlobalData\Client('127.0.0.1:55553');	 // initialize the GlobalData client
+	if (!isset($global->settings))
+		$global->settings = $settings;
 	if($worker->id === 0) { // The timer is set only on the process whose id number is 0, and the processes of other 1, 2, and 3 processes do not set the timer
-		Timer::add(600, 'update_vps_list_timer');
-		Timer::add(60, 'vps_queue_timer');
+		Timer::add($global->settings['timers']['update_vps_list'], 'update_vps_list_timer');
+		Timer::add($global->settings['timers']['vps_queue'], 'vps_queue_timer');
 	}
-
-	/*
-	Timer::add(60, function() use ($worker){
-		$time_now = time();
-		foreach ($worker->connections as $connection) {
-			// It is possible that the connection has not received the message, then lastMessageTime set to the current time
-			if (empty($connection->lastMessageTime)) {
-				$connection->lastMessageTime = $time_now;
-				continue;
+	if ($global->settings['heartbeat']['enable'] === TRUE) {
+		Timer::add($global->settings['heartbeat']['check_interval'], function() use ($worker, $global) {
+			$time_now = time();
+			foreach ($worker->connections as $connection) {
+				// It is possible that the connection has not received the message, then lastMessageTime set to the current time
+				if (empty($connection->lastMessageTime)) {
+					$connection->lastMessageTime = $time_now;
+					continue;
+				}
+				if ($time_now - $connection->lastMessageTime > $global->settings['heartbeat']['timeout'])
+					$connection->close();
 			}
-			if ($time_now - $connection->lastMessageTime > HEARTBEAT_TIME)
-				$connection->close();
+		});
+	}
+	if ($global->settings['vmstat']['enable'] === TRUE) {
+		// Save the process handle, close the handle when the process is closed
+		$worker->process_handle = popen('vmstat 1', 'r');
+		if ($worker->process_handle) {
+			$process_connection = new TcpConnection($worker->process_handle);
+			$process_connection->onMessage = function($process_connection, $data) use ($worker) {
+				foreach($worker->connections as $connection) {
+					$connection->send('vmstat:'.$data);
+				}
+			};
+		} else {
+		   echo "vmstat 1 fail\n";
 		}
-	});*/
+	}
 	/*
 	// async wss connection to my
 	$context_option = [
@@ -104,64 +139,53 @@ $worker->onWorkerStart = function($worker) {
 	};
 	$ws_connection->connect();
 	*/
-	// Save the process handle, close the handle when the process is closed
-	$worker->process_handle = popen('vmstat 1', 'r');
-	if ($worker->process_handle) {
-		$process_connection = new TcpConnection($worker->process_handle);
-		$process_connection->onMessage = function($process_connection, $data) use ($worker) {
-			foreach($worker->connections as $connection) {
-				$connection->send('vmstat:'.$data);
-			}
-		};
-	} else {
-	   echo "vmstat 1 fail\n";
-	}
 };
 
 $worker->onConnect = function($connection) {
-	/*
-	$connection->auth_timer_id = Timer::add(30, function()use($connection){
-		$connection->close();
-	}, null, false);
-	*/
-	// vmstat
-	$connection->send("vmstat:procs -----------memory---------- ---swap-- -----io---- -system-- ----cpu----\n");
-	$connection->send("vmstat:r  b   swpd   free   buff  cache   si   so    bi    bo   in   cs us sy id wa\n");
-	// phptty
-	//To do this, PHP_CAN_DO_PTS must be enabled. See ext/standard/proc_open.c in PHP directory.
-	/*$descriptorspec = [
-		0 => array('pty'),
-		1 => array('pty'),
-		2 => array('pty')
-	];*/
-	/*
-	//Pipe can not do PTY. Thus, many features of PTY can not be used. e.g. sudo, w3m, luit, all C programs using termios.h, etc.
-	$descriptorspec = [
-		0=>array("pipe", "r"),
-		1=>array("pipe", "w"),
-		2=>array("pipe", "w")
-	];
-	unset($_SERVER['argv']);
-	$env = array_merge(array('COLUMNS'=>130, 'LINES'=> 50), $_SERVER);
-	$connection->process = proc_open(CMD, $descriptorspec, $pipes, null, $env);
-	$connection->pipes = $pipes;
-	stream_set_blocking($pipes[0], 0);
-	$connection->process_stdout = new TcpConnection($pipes[1]);
-	$connection->process_stdout->onMessage = function($process_connection, $data)use($connection) {
-		$connection->send('phptty:'.$data);
-	};
-	$connection->process_stdout->onClose = function($process_connection)use($connection) {
-		$connection->close();   //Close WebSocket connection on process exit.
-	};
-	$connection->process_stdin = new TcpConnection($pipes[2]);
-	$connection->process_stdin->onMessage = function($process_connection, $data)use($connection) {
-		$connection->send('phptty:'.$data);
-	};
-	*/
+	global $global;
+	if ($global->settings['auth']['enable'] === TRUE) {
+		$connection->auth_timer_id = Timer::add(30, function() use ($connection){
+			$connection->close();
+		}, null, false);
+	}
+	if ($global->settings['vmstat']['enable'] === TRUE) {
+		$connection->send("vmstat:procs -----------memory---------- ---swap-- -----io---- -system-- ----cpu----\n");
+		$connection->send("vmstat:r  b   swpd   free   buff  cache   si   so    bi    bo   in   cs us sy id wa\n");
+	}
+	if ($global->settings['phptty']['enable'] === TRUE) {
+		//To do this, PHP_CAN_DO_PTS must be enabled. See ext/standard/proc_open.c in PHP directory.
+		/*$descriptorspec = [
+			0 => array('pty'),
+			1 => array('pty'),
+			2 => array('pty')
+		];*/
+		//Pipe can not do PTY. Thus, many features of PTY can not be used. e.g. sudo, w3m, luit, all C programs using termios.h, etc.
+		$descriptorspec = [
+			0=>array("pipe", "r"),
+			1=>array("pipe", "w"),
+			2=>array("pipe", "w")
+		];
+		unset($_SERVER['argv']);
+		$env = array_merge(array('COLUMNS'=>130, 'LINES'=> 50), $_SERVER);
+		$connection->process = proc_open($global->settings['phptty']['cmd'], $descriptorspec, $pipes, null, $env);
+		$connection->pipes = $pipes;
+		stream_set_blocking($pipes[0], 0);
+		$connection->process_stdout = new TcpConnection($pipes[1]);
+		$connection->process_stdout->onMessage = function($process_connection, $data) use ($connection) {
+			$connection->send('phptty:'.$data);
+		};
+		$connection->process_stdout->onClose = function($process_connection) use ($connection) {
+			$connection->close(); // Close WebSocket connection on process exit.
+		};
+		$connection->process_stdin = new TcpConnection($pipes[2]);
+		$connection->process_stdin->onMessage = function($process_connection, $data) use ($connection) {
+			$connection->send('phptty:'.$data);
+		};
+	}
 };
 
-
 $worker->onMessage = function($connection, $data) {
+	global $global;
 	$connection->lastMessageTime = time();
 	$data = json_decode($data, true);
 	switch ($data['type']) {
@@ -170,33 +194,35 @@ $worker->onMessage = function($connection, $data) {
 			Timer::del($connection->auth_timer_id);
 			break;
 		case 'phptty':
-			if(ALLOW_CLIENT_INPUT)
+			if ($global->settings['phptty']['client_input'] === TRUE)
 				fwrite($connection->pipes[0], $data);
 			break;
 	}
 };
 
 $worker->onClose = function($connection) {
-	/*
-	// phptty
-	$connection->process_stdin->close();
-	$connection->process_stdout->close();
-	fclose($connection->pipes[0]);
-	$connection->pipes = null;
-	proc_terminate($connection->process);
-	proc_close($connection->process);
-	$connection->process = null;
-	*/
+	global $global;
+	if ($global->settings['phptty']['enable'] === TRUE) {
+		$connection->process_stdin->close();
+		$connection->process_stdout->close();
+		fclose($connection->pipes[0]);
+		$connection->pipes = null;
+		proc_terminate($connection->process);
+		proc_close($connection->process);
+		$connection->process = null;
+	}
 };
 
 $worker->onWorkerStop = function($worker) {
-	// phptty
-	foreach($worker->connections as $connection) {
-		$connection->close();
+	global $global;
+	if ($global->settings['phptty']['enable'] === TRUE) {
+		foreach($worker->connections as $connection)
+			$connection->close();
 	}
-	// vmstat
-	@shell_exec('killall vmstat');
-	@pclose($worker->process_handle);
+	if ($global->settings['vmstat']['enable'] === TRUE) {
+		@shell_exec('killall vmstat');
+		@pclose($worker->process_handle);
+	}
 };
 
 $web = new WebServer("http://0.0.0.0:55555"); // WebServer, used to split html js css browser
